requirements.txt
streamlit==1.29.0
pandas==2.1.4
requests==2.31.0
google-generativeai==0.3.2
python-dotenv==1.0.0
plotly==5.17.0
jira==3.5.2
streamlit-authenticator==0.2.3
sqlite3